requirements.txt
pandas
scikit-learn
joblib
numpy
pyyaml